Fixing issue with resource authorization delete; Add acceptance tests (#322)
This change fixes an issue with the authorization delete. There was a
check that asserted that all updates to resource authorization would
result in at least one authorized resource being returned from the AzDO
API. This is not always the case. For example, a DELETE operation for
the last authorized resource will result in no objects being returned
from the API. This is not well documented in the API.