Make link text correspond to resource type name
Sorry for the nit but was slightly confused by not being able to find `azuredevops_git_repository` in the sidebar.  After clicking on `azuredevops_azure_git_repository` I found what I was looking for but this change should reduce friction for others.  I might suggest changing the linked pages names too but that's stylistic.