fix: flatten project falls back to organization default template ID
The Azure DevOps API does not guarantee a process template ID will
be returned when getting a project detail. This seems to be caused
for projects which use the organization default template. As such,
the lookup logic should fallback to that value, if none is present
at the project level. Thanks @tmeckel for the suggested
implementation.

Fixes #199