Initial resource providers for work item queries.
These include providers for:

- A work item query folder
- A work item query

We choose to split these up into seperate resources to make it simpler and easier to use. This is dispite the AzDO REST API using the same APIs for both folders and queries. If we were to use a single resource for both, we'd have to expose 'isFolder', which is a little messy and prone to misconfiguration.

This implementation is fairly basic, but in testing, and in the REST API examples, it seems only the wiql field is read when creating and updating. A lot of the other fields seem to be read-only. In testing, the vast majority of queries can be supported, including flat, linked and recursive.

Both the folder and query resource return the query or folder path in response. This allows a folder's path output to be used as the parent path for the query.

Feedback welcome.