`GetPrincipalPermissions` handle no permissions retrieved
`GetPrincipalPermissions()` is called by a few resources during `Read()`. One of the main step it does it to read the *identities* based on the principals (e.g. a user, a team, etc.) specified by the resource. Previously, the identity read logic is quite strict, where it expects the identity read matches the principal specified, with respect to the amount.

In most cases, this is fine. There is an edge case where the ADO org is changed (with the state records the original org), e.g. https://github.com/microsoft/terraform-provider-azuredevops/issues/1437. In this case, the identities read is `[null]`, which will be unmarshalled into an array of an empty identity. This causes a panic later on. To resolve this panic, we shall skip the empty identity, however, this is not enough since the existing identity read logic is strict about the amount,  where it expects the identity list returned can't be empty.

Instead, we shall be *honest* here to return whatever we read, in this case it is an empty list of identity. This allows the outside logic to attest the emptiness of the permission and set the resource as gone accordingly. For this reason, this PR introduces a function `tryGetIdentitiesFromSubjects`, which is similar to `getIdentitiesFromSubjects`, but more forgivable. This function replaces the original one, only in the `GetPrincipalPermissions`'s call site, which is used in the `Read()`. But keep the other call sites in `RemovePrincipalPermissions` and `SetPrincipalPermissions` untouched, to avoid unnecessary changes.