Addressing PR comments
We choose to use an area field to allow queries and folders to be placed into one of the immutable root folders. Either "Shared Queries" or "My Queries". This allows us to avoid using a data resource to find the ID of those folders.

We also choose to remove the path, as using this in queries results in issues where a query name is updated, the path isn't recalculated at plan time, resulting in the permissions applied to a query being incorrectly orphaned. Instead, the docs will show the use of the name attribute to create a path for the query permissions resource, which will result in the correct update behaviour.

Tests added, acceptance tests and docs to come.